Rename copied primer files to mention baseline
